feat(worker): add file-based sync logging and historical odds support
- Add SyncFileLogger singleton that writes per-date log files (sync-YYYY-MM-DD.log)
  and a failed-days.log for any day ending Partial or Failed
- Log every DB write (Game, GameLine, GameResult) with OK/FAIL entries
- Log sync lifecycle: SYNC_START and SYNC_COMPLETE with status and error count
- Register SyncFileLogger in DI; inject into SyncOrchestrator
- Add SyncLog:Path config (defaults to /logs); mount worker_logs volume in docker-compose
- Wire OddsApiHistoricalResponse and related OddsApi model/client changes

+++ nba-lines-tracker/src/NbaTracker.Worker/Models/OddsApi/OddsApiBookmaker.cs
@@ -18,6 +18,6 @@ public class OddsApiMarket
 public class OddsApiOutcome
 {
     [JsonPropertyName("name")] public string Name { get; set; } = null!;     // team name or "Over"/"Under"
-    [JsonPropertyName("price")] public int Price { get; set; }               // American odds: -110
+    [JsonPropertyName("price")] public decimal Price { get; set; }           // American odds: -110 (API returns as float)
     [JsonPropertyName("point")] public decimal? Point { get; set; }          // spread: -7.5 fav; total: 220.5
 }
